POCs/veronica/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
dill==0.3.4
distro==1.6.0
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMuPDF==1.19.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tabula-py==2.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.5.0
termcolor==1.1.0
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
